backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==6.0.2
Flask-SocketIO==5.6.0
Authlib==1.6.6
PyJWT==2.11.0
Werkzeug==3.1.3
Jinja2==3.1.6
requests==2.32.5
python-socketio==5.16.0
python-engineio==4.13.0
simple-websocket==1.1.0
cryptography==46.0.2
